common/src: move more implementation to C++
Converts proj_menu, software_cfu and cfu.h to C++ files.

Update proj_template to match. This commit breaks other projects, and
those will be fixed in future commits. proj_template is being committed
separately as it shows the scope of changes required.
